requirements.txt
# Core Flask (legacy, maintained for backwards compatibility)
Flask==3.1.0

# Core FastAPI (new API)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
starlette>=0.41.0

# Configuration
python-dotenv==1.1.0
PyYAML==6.0.2
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
neo4j==5.28.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12

# MCP Server
mcp>=1.0.0

# File handling
aiofiles>=24.1.0
python-magic>=0.4.27

# HTTP client (for MCP and external APIs)
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Development
ruff>=0.8.0
mypy>=1.13.0
